requirements.txt
opencv-python
pygame
numpy
django
djangorestframework
channels
asgiref